Only modify tmp basedir if new location is writable.
For some reason, some servers we're using create a doubly-nested LSF
tmpdir structure, and the inner one is r/w only by root.  In this case
they're already giving us the first level as $ENV{TMPDIR} so we can just
trust it anyway.  This code is still useful on our older cluster,
though, so not just removing it outright yet.